+++ lib/src/reporters/reporter.dart
@@ -1,0 +1,7 @@
+import 'dart:io';
+
+import '../models/analysis_report.dart';
+
+abstract class Reporter {
+  void report(AnalysisReport report, IOSink sink);
+}